cleanup

+++ src/gui/channel/autojoinchanneldialog.cpp
@@ -44,7 +44,7 @@ AutojoinChannelDialog::AutojoinChannelDialog(wxWindow* parent)
 	SetSizer(mainSizer);
 	wxString channels;
 	std::vector<ChannelJoinInfo> chanlist = sett().GetChannelsJoin();
-	for (int i = 0; i < long(chanlist.size()); i++) {
+	for (size_t i = 0; i < chanlist.size(); i++) {
 		ChannelJoinInfo info = chanlist[i];
 		channels += info.name;
 		if (!info.password.empty()) {
